Add New Site: Unify the Add New Site modal between new and v2 dashboard
