+++ client/components/sites-add-new-site/async.tsx
@@ -2,10 +2,5 @@ import { Spinner } from '@wordpress/components';
 import AsyncLoad from 'calypso/components/async-load';
 
 export const AsyncContent = () => {
-	return (
-		<AsyncLoad
-			require="calypso/components/sites-add-new-site/content"
-			placeholder={ <Spinner /> }
-		/>
-	);
+	return <AsyncLoad require="calypso/dashboard/sites/add-new-site" placeholder={ <Spinner /> } />;
 };
