--- client/components/sites-add-new-site/content/layout/column.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-type Props = {
-	heading?: string;
-	children: React.ReactNode;
-};
-
-export const Column = ( { heading, children }: Props ) => {
-	return (
-		<div className="sites-add-new-site__popover-column">
-			{ heading && (
-				<div className="sites-add-new-site__popover-column-heading">{ heading.toUpperCase() }</div>
-			) }
-			{ children }
-		</div>
-	);
-};
