+++ client/components/sites-add-new-site/style.scss
@@ -22,149 +22,10 @@
 .sites-add-new-site__popover-content {
 	display: flex;
 	flex-direction: column;
-	padding: 1em;
 	gap: 1em;
 
 	@include break-large {
 		flex-direction: row;
-		padding: 28px 32px;
 		gap: 32px;
-	}
-
-	.sites-add-new-site__popover-column {
-		width: 260px;
-		display: flex;
-		flex-direction: column;
-		gap: 8px;
-	}
-}
-
-.sites-add-new-site__popover-column-heading {
-	color: var(--color-neutral-40);
-	font-size: rem(12px);
-	font-weight: 500;
-	padding: 0 1em;
-
-	@include break-large {
-		padding: 0;
-	}
-}
-
-
-.components-button.sites-add-new-site__popover-button {
-	white-space: unset;
-	text-align: unset;
-	margin-block: 0;
-	width: 100%;
-	border-radius: 4px;
-	height: auto;
-	color: var(--color-text-black);
-	padding: 1em;
-
-	&:hover {
-		color: var(--color-text-black);
-		background-color: var(--color-neutral-0);
-	}
-
-	&.is-disabled {
-		cursor: not-allowed;
-		background-color: var(--color-neutral-0);
-	}
-
-	@include break-large {
-		padding: 8px;
-		display: flex;
-		gap: 8px;
-		align-items: unset;
-		justify-content: unset;
-	}
-
-	&.is-banner {
-		flex-direction: column;
-		gap: 4px;
-		background-color: var(--color-neutral-0);
-
-		@include break-large {
-			padding: 16px;
-		}
-
-		.sites-add-new-site__popover-button-icon {
-			height: 78px;
-
-			.sidebar__menu-icon {
-				width: 228px;
-				height: auto;
-			}
-		}
-
-		.sites-add-new-site__popover-button-content {
-			@include break-large {
-				padding: 8px 8px 0 8px;
-			}
-		}
-
-		.sites-add-new-site__popover-button-heading {
-			font-size: rem(14px);
-			font-weight: 500;
-		}
-	}
-
-	.sites-add-new-site__popover-button-icon {
-		display: none;
-
-		@include break-large {
-			display: block;
-		}
-
-		.sidebar__menu-icon {
-			width: 24px;
-			height: 24px;
-			max-width: fit-content;
-			color: currentColor;
-		}
-	}
-
-	.sites-add-new-site__popover-button-heading {
-		font-size: rem(14px);
-		font-weight: 600;
-		text-wrap: balance; /* Fallback for browsers that don't support pretty. */
-		text-wrap: pretty;
-	}
-
-	.sites-add-new-site__popover-button-description {
-		font-size: rem(12px);
-		font-weight: 400;
-		color: var(--color-text-black);
-		padding-top: 4px;
-		text-wrap: balance; /* Fallback for browsers that don't support pretty. */
-		text-wrap: pretty;
-	}
-}
-
-// Remove the focus outline from the button when clicked with the mouse,
-// but retain it for keyboard navigation (e.g., Tab key). */
-html:not(.accessible-focus) button,
-a {
-	&.components-button.sites-add-new-site__popover-button {
-		&:focus-within {
-			box-shadow: none;
-		}
-	}
-}
-
-.sites-add-new-site-popover__count {
-	font-size: rem(12px);
-	font-style: italic;
-	color: var(--color-accent-100);
-	padding-top: 4px;
-}
-
-.sites-add-new-site-popover__cta {
-	padding-top: 12px;
-	font-size: rem(12px);
-	font-weight: 600;
-
-	&.disabled {
-		color: var(--color-accent-20);
 	}
 }
