+++ client/dashboard/sites/add-new-site/style.scss
@@ -7,9 +7,9 @@
 	height: auto;
 	text-align: left;;
 	padding: 8px;
+	color: inherit !important; // avoid changing the color on hover
 
 	&:hover {
-		color: inherit !important; // avoid changing the color on hover
 		background: $gray-100;
 		border-radius: $radius-medium;
 	}
